requirements.txt
langchain 
langchain-openai
psycopg2-binary
pgvector
langchain-community
langchain-chroma
langchain-deepseek
dotenv
requests
beautifulsoup4
sentence-transformers
cohere
numpy
ragas
datasets